requirements.txt
Flask==0.10.1
Jinja2==2.7.1

MarkupSafe==0.18

Werkzeug==0.9.4

blinker==1.3

certifi==0.0.8

chardet==2.2.1

cssselect==0.6.1
pyquery==1.2.4


itsdangerous==0.23

lxml==3.2.4

python-dateutil==2.2

pytz
raven==3.5.0

requests==2.1.0

simplejson==3.3.1

six==1.4.1

Twisted==13.2.0
flask-sqlalchemy==1.0
psycopg2==2.5
APScheduler==2.0.3